pyproject.toml
[tool.poetry]
name = "imagecap"
version = "0.1.0"
description = ""
authors = ["michaelbahchevanov <bahchevanov.mihail@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
addict = "2.4.0"
annotated-types = "0.6.0"
anyio = "4.2.0"
appnope = "0.1.3"
argon2-cffi = "23.1.0"
argon2-cffi-bindings = "21.2.0"
arrow = "1.3.0"
asttokens = "2.4.1"
async-lru = "2.0.4"
attrs = "23.2.0"
babel = "2.14.0"
beautifulsoup4 = "4.12.3"
bleach = "6.1.0"
certifi = "2023.11.17"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
comm = "0.2.1"
contourpy = "1.2.0"
cycler = "0.12.1"
debugpy = "1.8.0"
decorator = "5.1.1"
defusedxml = "0.7.1"
exceptiongroup = "1.2.0"
executing = "2.0.1"
fastjsonschema = "2.19.1"
filelock = "3.13.1"
fonttools = "4.47.2"
fqdn = "1.5.1"
fsspec = "2023.12.2"
groundingdino-py = "0.4.0"
huggingface-hub = "0.20.3"
idna = "3.6"
importlib-metadata = "7.0.1"
ipykernel = "6.29.0"
ipython = "8.21.0"
ipywidgets = "8.1.1"
isoduration = "20.11.0"
jedi = "0.19.1"
jinja2 = "3.1.3"
json5 = "0.9.14"
jsonpointer = "2.4"
jsonschema = "4.21.1"
jsonschema-specifications = "2023.12.1"
jupyter = "1.0.0"
jupyter-console = "6.6.3"
jupyter-events = "0.9.0"
jupyter-lsp = "2.2.2"
jupyter-client = "8.6.0"
jupyter-core = "5.7.1"
jupyter-server = "2.12.5"
jupyter-server-terminals = "0.5.2"
jupyterlab = "4.0.12"
jupyterlab-widgets = "3.0.9"
jupyterlab-pygments = "0.3.0"
jupyterlab-server = "2.25.2"
kiwisolver = "1.4.5"
markupsafe = "2.1.4"
matplotlib = "3.8.2"
matplotlib-inline = "0.1.6"
mistune = "3.0.2"
mpmath = "1.3.0"
nbclient = "0.9.0"
nbconvert = "7.14.2"
nbformat = "5.9.2"
nest-asyncio = "1.6.0"
networkx = "3.2.1"
notebook = "7.0.7"
notebook-shim = "0.2.3"
numpy = "1.26.3"
opencv-python = "4.9.0.80"
opencv-python-headless = "4.9.0.80"
overrides = "7.7.0"
packaging = "23.2"
pandas = "2.2.0"
pandocfilters = "1.5.1"
parso = "0.8.3"
pexpect = "4.9.0"
pillow = "10.2.0"
platformdirs = "4.2.0"
prometheus-client = "0.19.0"
prompt-toolkit = "3.0.43"
psutil = "5.9.8"
ptyprocess = "0.7.0"
pure-eval = "0.2.2"
py-cpuinfo = "9.0.0"
pycocotools = "2.0.7"
pycparser = "2.21"
pydantic = "2.6.0"
pydantic-core = "2.16.1"
pygments = "2.17.2"
pyparsing = "3.1.1"
python-dateutil = "2.8.2"
python-json-logger = "2.0.7"
pytz = "2023.4"
pyyaml = "6.0.1"
pyzmq = "25.1.2"
qtconsole = "5.5.1"
qtpy = "2.4.1"
referencing = "0.33.0"
regex = "2023.12.25"
requests = "2.31.0"
rfc3339-validator = "0.1.4"
rfc3986-validator = "0.1.1"
rpds-py = "0.17.1"
safetensors = "0.4.2"
scipy = "1.12.0"
seaborn = "0.13.2"
segment-anything = "1.0"
send2trash = "1.8.2"
six = "1.16.0"
sniffio = "1.3.0"
soupsieve = "2.5"
stack-data = "0.6.3"
supervision = "0.6.0"
sympy = "1.12"
terminado = "0.18.0"
thop = "0.1.1.post2209072238"
timm = "0.9.12"
tinycss2 = "1.2.1"
tokenizers = "0.15.1"
tomli = "2.0.1"
torch = "2.2.0"
torchvision = "0.17.0"
tornado = "6.4"
tqdm = "4.66.1"
traitlets = "5.14.1"
transformers = "4.37.2"
types-python-dateutil = "2.8.19.20240106"
typing-extensions = "4.9.0"
tzdata = "2023.4"
ultralytics = "8.1.8"
uri-template = "1.3.0"
urllib3 = "2.2.0"
wcwidth = "0.2.13"
webcolors = "1.13"
webencodings = "0.5.1"
websocket-client = "1.7.0"
widgetsnbextension = "4.0.9"
yapf = "0.40.2"
zipp = "3.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
extend-select = ['Q', 'RUF100', 'C90', 'UP', 'I', 'D', 'T']
extend-ignore = ['D105', 'D107', 'D205', 'D415']
extend-exclude = ["**/*.json",
                  "exploration",
                  "docker",
                  "Makefile",
                  ]
src = ["src/imagecap"]
force-exclude = true
mccabe = { max-complexity = 14 }

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401", "F403"]
